requirements.txt
jupyterlab
pandas
boto3
pymongo
kafka-python